feat: fetches changes since last gh release and thus emoving from flag
The reason for that is following:

- from is not feasible for automated release, as at this point, when
triggered by a tag it only knows current release tag. Therefore lookup
for previous tag is required which turns out to be quite cumbersome with
gh graphql. Querying with cursore and (first: 1, before: "cursor", ...)
cannot be based on gh oid (hash), even though query is not invalid
(cursor is accepted). But it always returns first ever tags. Technically
we could retrieve all tags (or all releases) and just iterate to find
previous tag, but that seems to be an overhead and fits only marginal
usecases.
- most of the times usage of the tool will be for the latest release, so
no need for lookup previous tags, but last release

Not working yet - first use for the first release - no releases on GH
yet. This will currently fail - we have to lookup the very first commit