feat(deps): include all "bumps" dep exp in commits
Scan commit message for "bumps" version dependency expressions
anywhere. This allow "bumps" in regular feature commits to be
included in the deps updated section.

Scan each line in all commit messages and clone the PR object if one
found and append to the deps PR list. PRs do not have to include a
"dependencies" label to be included in the scan. If it does contain
a "dependencies" label it will be excluded from the "others" PR list
and not show up as a "feature".